relocated fetch image method to auth service
