+++ backend/src/auth/dto/update.dto.ts
@@ -1,5 +1,5 @@
 import { ApiProperty } from '@nestjs/swagger';
-import { IsOptional, IsString } from 'class-validator';
+import { IsOptional, IsString, Matches, MinLength } from 'class-validator';
 
 export class UpdateDto {
   @ApiProperty({ description: 'New username (optional)' })
@@ -10,6 +10,8 @@ export class UpdateDto {
   @ApiProperty({ description: 'New password (optional)' })
   @IsString()
   @IsOptional()
+  @MinLength(8)
+  @Matches(/((?=.*\d)|(?=.*\W+))(?![.\n])(?=.*[A-Z])(?=.*[a-z]).*$/, {message: 'password too weak. Need at least one uppercase letter, one lowercase letter and one digit.'})
   password: string;
 
   @ApiProperty({
