+++ backend/src/auth/strategy/FourtyTwo.strategy.ts
@@ -7,7 +7,6 @@ import { lastValueFrom } from 'rxjs';
 // for env
 import { ConfigService } from '@nestjs/config';
 import { AuthDto } from '../dto';
-import axios from 'axios';
 
 @Injectable()
 export class FourtyTwoStrategy extends PassportStrategy(Strategy, '42') {
@@ -50,25 +49,25 @@ export class FourtyTwoStrategy extends PassportStrategy(Strategy, '42') {
     user.id = userData.id.toString();
     user.userName = userData.login;
     user.email = userData.email;
-    user.image = await this.fetchImage(userData.image.link);
+    user.image = userData.image.link;
     user.password = '';
 
     return cb(null, user, accessToken);
   }
 
-  async fetchImage(url: string): Promise<string> {
-    try {
-      const response = await axios.get(url, {
-        responseType: 'arraybuffer',
-      });
-      if (response.status === 200) {
-        const imageBuffer = Buffer.from(response.data, 'binary');
-        const returnString = imageBuffer.toString('base64');
-        return returnString;
-      }
-    } catch (error) {
-      throw new NotFoundException('Could not load profile picture.');
-    }
-    return null;
-  }
+//   async fetchImage(url: string): Promise<string> {
+//     try {
+//       const response = await axios.get(url, {
+//         responseType: 'arraybuffer',
+//       });
+//       if (response.status === 200) {
+//         const imageBuffer = Buffer.from(response.data, 'binary');
+//         const returnString = imageBuffer.toString('base64');
+//         return returnString;
+//       }
+//     } catch (error) {
+//       throw new NotFoundException('Could not load profile picture.');
+//     }
+//     return null;
+//   }
 }
