+++ backend/src/user/user.service.ts
@@ -48,6 +48,7 @@ export class UserService {
       'PongStoryShort',
       secret,
     );
+	// redirect to page with otpautUrl + route '/2fa-verify'
     return otpauthUrl;
   }
 }
